feat: add TaskSearchHeader
